+++ cmd/main.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 	"github.com/emyt-io/emyt/config"
+	"github.com/emyt-io/emyt/config/models"
 	manager "github.com/emyt-io/emyt/manager"
-	"github.com/emyt-io/emyt/models"
 	"github.com/ilyakaznacheev/cleanenv"
 	"github.com/labstack/echo/v4"
 	"github.com/labstack/echo/v4/middleware"
