+++ db/database.go
@@ -1,0 +1,7 @@
+package db
+
+import "github.com/jinzhu/gorm"
+
+type DbOps struct {
+	db *gorm.DB
+}
