+++ db/models/user.go
@@ -1,4 +1,4 @@
-package manager
+package models
 
 import "github.com/jinzhu/gorm"
 
@@ -6,8 +6,4 @@ type User struct {
 	gorm.Model `json:"model"`
 	Name       string `json:"name"`
 	Email      string `json:"email"`
-}
-
-type DbOps struct {
-	db *gorm.DB
 }
